+++ context_test.go
@@ -103,7 +103,7 @@ func ExampleContext_isolate() {
 func ExampleContext_globalTemplate() {
 	iso, _ := v8go.NewIsolate()
 	obj, _ := v8go.NewObjectTemplate(iso)
-	obj.Set("version", "v1.0.0", 0)
+	obj.Set("version", "v1.0.0")
 	ctx, _ := v8go.NewContext(iso, obj)
 	val, _ := ctx.RunScript("version", "main.js")
 	fmt.Println(val)